refactor: 같은파일에서 사용할수 없는 로직 분리
